JobStatus/Feature tracking, AtomicTT separation, strudel leak fixes
Always-on intrusive linked list tracking for JobStatus (Channel/LockBox)
and Feature objects. Shared ID counter, creation location, Feature-to-owner
relationship. DumpJobQueLeaks() at exit reports all surviving objects with
ownership graph. --track-job-status <id> CLI arg for per-object
addRef/releaseRef tracing.

AtomicTT no longer inherits from JobStatus -- it is a bare std::atomic
wrapper with no refcounting.

Fix LockBox leaks in strudel_player, strudel_midi_player: fill/grab/join
is sufficient sync, removed redundant add_ref/release. Fix main-thread
mode leak: strudel_shutdown early-returned when g_cmd_channel was null,
skipping g_tick_status_box cleanup.

Add proper shutdown to strudel_live, strudel_sf2_live, strudel_visualizer
(were missing strudel_shutdown + audio_system_finalize).

Fix daslang-live ModuleGroup crash on shutdown (was stored in CompileResult,
destroyed after modules were gone).

Add --max-frames to arcanoid and strudel_visualizer for automated testing.
Add jobque_debugging skill file.

+++ src/ast/ast_infer_type_op.cpp
@@ -640,7 +640,7 @@ namespace das {
                         auto fidx = efield->value->type->bitFieldIndex(efield->name);
                         if (fidx != -1) {
                             reportAstChanged();
-                            auto mask = new ExprConstBitfield(efield->at, 1ul << fidx);
+                            auto mask = new ExprConstBitfield(efield->at, uint64_t(1) << uint64_t(fidx));
                             mask->bitfieldType = new TypeDecl(*efield->value->type);
                             auto call = new ExprCall(efield->at, "__bit_set");
                             call->arguments.push_back(value->clone());
